# libs/domains/agenstra/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
  - url: http://localhost:3100
    description: Root (OpenTelemetry metrics; outside /api prefix)
security:
  - bearerAuth: []
paths:
...
  /clients/{id}:
    get:
      summary: Get client by id
      description: |
        Returns a client by ID. Only accessible if the user has access to the client.
        Access is granted if:
        - The user is a global admin
        - The user is the creator (user_id matches)
        - The user is in the client_users relationship table
        In api-key mode, access is always granted.
      operationId: getClient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponseDto'
        '403':
          description: User does not have access to this client
    post:
      summary: Update client
      description: |
        Updates a client. Requires workspace management rights: global admin, workspace creator,
        client_users role `admin`, API key authentication, or api-key server mode (not plain client_users `user`).
        See `canManageWorkspaceConfiguration` on ClientResponseDto.
      operationId: updateClient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientDto'
      responses:
        '200':
          description: Updated client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponseDto'
        '403':
          description: |
            Forbidden — no access to this client, or caller lacks workspace management rights
            (same rule as POST for updates).
    delete:
      summary: Delete client
      description: |
        Deletes a client. Same workspace management requirement as POST /clients/{id} (not available to plain client_users `user`).
      operationId: deleteClient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '403':
          description: |
            Forbidden — no access to this client, or caller lacks workspace management rights.
...
  /clients/{id}/opencode/config:
    get:
      summary: Get workspace OpenCode config
      operationId: getClientOpencodeConfig
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workspace OpenCode config overlay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeConfigResponseDto'
    put:
      summary: Update workspace OpenCode config
      operationId: putClientOpencodeConfig
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertOpencodeConfigDto'
      responses:
        '200':
          description: Updated workspace OpenCode config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeConfigResponseDto'
...
  /clients/{id}/agents/{agentId}/opencode/config:
    get:
      summary: Get agent OpenCode config
      operationId: getClientAgentOpencodeConfig
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent OpenCode config with effective merge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeConfigResponseDto'
    put:
      summary: Update agent OpenCode config
      operationId: putClientAgentOpencodeConfig
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertOpencodeConfigDto'
      responses:
        '200':
          description: Updated agent OpenCode config with effective merge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeConfigResponseDto'
  /clients/{id}/agents/{agentId}/opencode/config/agents:
    get:
      summary: List effective OpenCode agents for a remote worker
      operationId: listClientAgentOpencodeAgents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Effective OpenCode agents (primary + subagents)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeAgentsListResponseDto'
  /clients/{id}/agents/{agentId}/opencode/config/commands:
    get:
      summary: List slash-invokable OpenCode commands for a remote worker
      description: >
        Proxies the agent-manager list of OpenCode `GET /command` entries (config commands,
        builtins, skills, MCP prompts).
      operationId: listClientAgentOpencodeCommands
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Slash commands available on the worker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeCommandsListResponseDto'
  /clients/{id}/agents/{agentId}/opencode/mcp:
    get:
      summary: List live MCP server statuses for an Environment
      description: >
        Environment-only. Proxies OpenCode `GET /mcp` via the agent manager. Use to detect
        servers that need interactive OAuth (`needs_auth` / `needs_client_registration`).
      operationId: listClientAgentOpencodeMcpStatuses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: MCP server statuses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeMcpStatusListResponseDto'
  /clients/{id}/agents/{agentId}/opencode/mcp/{name}/auth:
    post:
      summary: Start interactive MCP OAuth for an Environment MCP server
      description: >
        Environment-only. Builds a signed public callback URL, sets it as the worker
        oauth.redirectUri, then proxies OpenCode `POST /mcp/{name}/auth`.
      operationId: startClientAgentOpencodeMcpAuth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth flow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeMcpAuthStartResponseDto'
        '400':
          description: Agent not running, MCP OAuth unsupported, or callback env not configured
    delete:
      summary: Remove MCP OAuth credentials for an Environment MCP server
      operationId: removeClientAgentOpencodeMcpAuth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth credentials removed
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                    enum: [true]
  /clients/{id}/agents/{agentId}/opencode/mcp/{name}/auth/callback:
    post:
      summary: Complete MCP OAuth with authorization code
      operationId: completeClientAgentOpencodeMcpAuth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Updated MCP status after OAuth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeMcpServerStatusDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
    AuthenticationType:
      type: string
      enum:
        - api_key
        - keycloak
...
    ClientResponseDto:
      type: object
      required:
        [
          id,
          name,
          endpoint,
          authenticationType,
          isAutoProvisioned,
          canManageWorkspaceConfiguration,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
          format: uri
        authenticationType:
          $ref: '#/components/schemas/AuthenticationType'
        agentWsPort:
          type: integer
          minimum: 1
          maximum: 65535
          description: Optional Socket.IO port for the client's agents gateway
        config:
          $ref: '#/components/schemas/ConfigResponseDto'
          description: Configuration from the agent-manager instance (optional, may be undefined if agent-manager is unreachable)
        isAutoProvisioned:
          type: boolean
          description: Indicates whether the client was auto-provisioned via server provisioning. Used to determine if server info endpoints should be called.
        canManageWorkspaceConfiguration:
          type: boolean
          description: |
            True if the current caller may change prototype autonomy, agent environment variables, agent CRUD,
            and workspace (client) update/delete. Requires global admin, workspace creator, client_users admin,
            API key authentication, or api-key server mode.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateClientDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
          format: uri
        authenticationType:
          $ref: '#/components/schemas/AuthenticationType'
        apiKey:
          type: string
          description: API key (can be updated but will never be included in responses)
        keycloakClientId:
          type: string
          description: Keycloak client ID
        keycloakClientSecret:
          type: string
          description: Keycloak client secret
        keycloakRealm:
          type: string
          description: Keycloak realm
        agentWsPort:
          type: integer
          minimum: 1
          maximum: 65535
          description: Optional Socket.IO port for the client's agents gateway
...
    AgentTypeInfo:
      type: object
      required: [type, displayName, capabilities]
      properties:
        type:
          type: string
          description: The unique type identifier (e.g., 'opencode')
        displayName:
          type: string
          description: Human-readable display name (e.g., 'OpenCode')
        capabilities:
          $ref: '#/components/schemas/AgentTypeCapabilities'
    AgentTypeCapabilities:
      type: object
      required:
        [supportsChat, supportsStreaming, supportsToolEvents, supportsQuestions]
      properties:
        transport:
          type: string
          enum: [acp]
          description: Wire transport for agent messaging (Agent Client Protocol over stdio)
        supportsChat:
          type: boolean
        supportsStreaming:
          type: boolean
        supportsToolEvents:
          type: boolean
        supportsQuestions:
          type: boolean
    ConfigResponseDto:
      type: object
      required: [agentTypes]
      properties:
        gitRepositoryUrl:
          type: [string, 'null']
          description: Git repository URL used for cloning repositories when creating agents
        gitRepositorySetupMode:
          type: string
          enum: [clone, empty]
          description: Default setup mode for new agents on this instance
        agentTypes:
          type: array
          items:
            $ref: '#/components/schemas/AgentTypeInfo'
          description: Array of available agent provider types with display names
...
    OpencodeConfigResponseDto:
      type: object
      required: [config, overrides, secretKeys]
      properties:
        config:
          type: object
          additionalProperties: true
        overrides:
          type: object
          additionalProperties: true
          description: >
            Raw JSON overrides for this layer, merged over `config` before
            cross-layer merge (same heredity locks as `config`).
        secretKeys:
          type: array
          items:
            type: string
        effective:
          type: object
          additionalProperties: true
          description: >
            Full merged config for this scope (agent → workspace → global).
            Each layer is compose(config, overrides) before cross-layer merge.
            Platform keys such as model_allow / model_deny are retained for the editor;
            worker sync materializes them separately.
            The `config` / `overrides` fields remain the editable overlays for the current layer only.
        lockedPaths:
          type: array
          items:
            type: string
          description: JSON Pointer paths locked by a higher layer
        inheritedAdditive:
          type: array
          items:
            type: object
            required: [path]
            properties:
              path:
                type: string
              keys:
                type: array
                items:
                  type: string
              items:
                type: array
        updatedAt:
          type: string
          format: date-time
        sync:
          type: object
          nullable: true
          description: Durable per-agent effective-config sync target (agent GET/PUT)
          properties:
            syncStatus:
              type: string
              enum: [pending, synced, failed]
            desiredRevision:
              type: string
            appliedRevision:
              type: string
              nullable: true
            lastError:
              type: string
              nullable: true
            lastSyncedAt:
              type: string
              format: date-time
              nullable: true
...
    OpencodeAgentsListResponseDto:
      type: object
      required: [agents]
      properties:
        agents:
          type: array
          items:
            type: object
            required: [name, mode, builtIn]
            properties:
              name:
                type: string
              description:
                type: string
              mode:
                type: string
                enum: [subagent, primary, all]
              builtIn:
                type: boolean
              prompt:
                type: string
              temperature:
                type: number
              color:
                type: string
              model:
                type: object
                properties:
                  modelID:
                    type: string
                  providerID:
                    type: string
    OpencodeCommandsListResponseDto:
      type: object
      required: [commands]
      properties:
        commands:
          type: array
          items:
            type: object
            required: [name]
            properties:
              name:
                type: string
              description:
                type: string
              agent:
                type: string
              model:
                type: string
              subtask:
                type: boolean
              source:
                type: string
                enum: [command, mcp, skill]
    OpencodeMcpServerStatusDto:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum:
            [connected, disabled, failed, needs_auth, needs_client_registration]
        error:
          type: string
    OpencodeMcpStatusListResponseDto:
      type: object
      required: [servers]
      properties:
        servers:
          type: array
          items:
            $ref: '#/components/schemas/OpencodeMcpServerStatusDto'
    OpencodeMcpAuthStartResponseDto:
      type: object
      required: [authorizationUrl]
      properties:
        authorizationUrl:
          type: string
          format: uri
        oauthState:
          type: string
    UpsertOpencodeConfigDto:
      type: object
      properties:
        config:
          type: object
          nullable: true
          additionalProperties: true
        overrides:
          type: object
          nullable: true
          additionalProperties: true
          description: >
            Raw JSON patch merged over `config` for this layer.
            Same credential, v1-key, and heredity checks as `config`.
            Empty object clears overrides when provided.
        secrets:
          type: object
          nullable: true
          additionalProperties:
            type: string